requirements.txt
# Flask Web框架
Flask==3.0.0

# 环境变量管理
python-dotenv==1.0.0

# HTTP请求库
requests==2.31.0

# OpenAI SDK（用于阿里云百炼）
openai==1.12.0

# 系统监控（用于日志记录）
psutil==5.9.8

# 音频处理（可选）
# pydub==0.25.1

# 腾讯云SDK（可选，用于语音合成）
# tencentcloud-sdk-python==3.0.1133

# 数据库和认证
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
bcrypt==4.1.2

# MySQL连接
pymysql==1.1.0

# 开发工具
# black==23.11.0
# flake8==6.1.0
# pytest==7.4.3